fix: begin cleaning up token types
