+++ apps/namada-interface/src/App/Token/IBCTransfer/IBCTransfer.tsx
@@ -6,13 +6,13 @@ import {
   Account as AccountType,
   BridgeType,
   Chain,
-  CosmosTokens,
   CosmosTokenType,
   ExtensionKey,
   Extensions,
   TokenType,
   Tokens,
   ChainKey,
+  CosmosTokens,
 } from "@namada/types";
 import {
   Alert,
