+++ packages/types/src/index.ts
@@ -1,7 +1,8 @@
 export * from "./account";
-export * from "./namada";
 export * from "./chain";
 export * from "./events";
+export * from "./namada";
 export * from "./signer";
+export * from "./tokens";
 export * from "./tx";
 export * from "./utils";
