+++ packages/types/src/tx/index.ts
@@ -1,4 +1,3 @@
 export * from "./messages";
 export * from "./schema";
-export * from "./tokens";
 export * from "./types";
